third_party/blink/web_tests/external/wpt/css/cssom-view/window-scroll-promises-expected.txt

This is a testharness.js-based test.
[FAIL] Window.scroll(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
[FAIL] Window.scrollTo(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
[FAIL] Window.scrollBy(smooth): promise behavior
  assert_equals: Position after await expected 5000 but got 0
Harness: the test ran to completion.
